tests/requirements.txt
boto3>=1.26.0
cfn-lint>=0.72.0
taskcat>=0.9.0
pyyaml>=6.0
cfn_tools>=0.1.6
pytest>=7.0.0
moto>=4.0.0
